Added Authentication and Authorization using Auth0

+++ pages/api/snippet/[id].js
@@ -3,8 +3,11 @@ import {
 	getSnippetById,
 	updateSnippet,
 } from "../../../utils/Fauna";
+import { getSession, withApiAuthRequired } from "@auth0/nextjs-auth0";
 
-export default async function handler(req, res) {
+export default withApiAuthRequired(async function handler(req, res) {
+	const session = getSession(req, res);
+	const userId = session.user.sub;
 	switch (req.method) {
 		case "GET": {
 			try {
@@ -19,6 +22,12 @@ export default async function handler(req, res) {
 		case "PUT": {
 			const { code, language, description, name } = req.body;
 			try {
+				const existingRecord = await getSnippetById(req.query.id);
+				if (!existingRecord || existingRecord.data.userId !== userId) {
+					return res
+						.status(403)
+						.json({ msg: "Unauthorised. Access denied." });
+				}
 				const updatedSnippet = await updateSnippet(
 					req.query.id,
 					code,
@@ -26,6 +35,7 @@ export default async function handler(req, res) {
 					description,
 					name
 				);
+
 				return res.status(201).json(updatedSnippet);
 			} catch (error) {
 				console.error(err);
@@ -35,6 +45,12 @@ export default async function handler(req, res) {
 		}
 		case "DELETE": {
 			try {
+				const existingRecord = await getSnippetById(req.query.id);
+				if (!existingRecord || existingRecord.data.userId !== userId) {
+					return res
+						.status(403)
+						.json({ msg: "Unauthorised. Access denied." });
+				}
 				const deletedSnippet = await deleteSnippet(req.query.id);
 				return res.status(200).json(deletedSnippet);
 			} catch (error) {
@@ -46,4 +62,4 @@ export default async function handler(req, res) {
 		default:
 			return res.status(405).json({ msg: "Method not allowed" });
 	}
-}
+});
